@@ -2,21 +2,22 @@
 // Licensed under the MIT License.
 import { title } from 'content/strings/application';
 import * as React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import { NewTabLink } from '../common/components/new-tab-link';
 import { Header } from './components/header';
 
 export class IncompatibleBrowserRenderer {
     constructor(
-        private readonly renderer: typeof ReactDOM.render,
+        private readonly renderer: typeof createRoot,
         private readonly dom: Document,
-    ) {}
+    ) { }
 
     public render(): void {
         const container = this.dom.querySelector('#popup-container');
+        const root = this.renderer(container)
 
-        this.renderer(
+        root.render(
             <>
                 <div className="ms-Fabric unsupported-browser-info-panel">
                     <Header title={title} />
@@ -38,7 +39,6 @@ export class IncompatibleBrowserRenderer {
                     </div>
                 </div>
             </>,
-            container,
         );
     }
 }
