@@ -22,7 +22,8 @@ import { UserConfigurationStoreData } from 'common/types/store-data/user-configu
 import { contentPages } from 'content';
 import { textContent } from 'content/strings/text-content';
 import { ScannerUtils } from 'injected/scanner-utils';
-import ReactDOM from 'react-dom';
+
+import {createRoot} from 'react-dom/client';
 import { scan } from 'scanner/exposed-apis';
 import UAParser from 'ua-parser-js';
 import { Content } from 'views/content/content';
@@ -69,7 +70,7 @@ const documentManipulator = new DocumentManipulator(document);
 const rendererDependencies: RendererDeps = {
     textContent,
     dom: document,
-    render: ReactDOM.render,
+    render: createRoot,
     initializeFabricIcons,
     loadTheme,
     contentProvider: contentPages,
